# backend-nodejs/docs/api-documentation.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Business Intelligence API for University Admission Data",
    "version": "1.0.0",
    "description": "API for processing and analyzing university admission test data",
    "contact": {
      "name": "API Support",
      "email": "support@university.edu"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "AdmissionData": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Unique identifier for the admission record"
          },
          "period": {
            "type": "string",
            "description": "Academic period (e.g., 2025-2)"
          },
          "year": {
            "type": "number",
            "description": "Academic year"
          },
          "studentId": {
            "type": "string",
            "description": "Student identification number"
          },
          "studentName": {
            "type": "string",
            "description": "Full name of the student"
          },
          "scores": {
            "type": "object",
            "description": "Object containing scores for different test subjects",
            "additionalProperties": {
              "type": "number"
            }
          },
          "totalScore": {
            "type": "number",
            "description": "Total score across all subjects"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "approved",
              "rejected"
            ],
            "description": "Admission status based on scores"
          },
          "uploadedBy": {
            "type": "string",
            "description": "ID of the user who uploaded this data"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the record was created"
          }
        }
      },
      "AdmissionStats": {
        "type": "object",
        "properties": {
          "totalStudents": {
            "type": "number",
            "description": "Total number of students"
          },
          "averageScore": {
            "type": "number",
            "description": "Average score across all students"
          },
          "maxScore": {
            "type": "number",
            "description": "Highest score achieved"
          },
          "minScore": {
            "type": "number",
            "description": "Lowest score achieved"
          },
          "approvedCount": {
            "type": "number",
            "description": "Number of students who were approved"
          },
          "rejectedCount": {
            "type": "number",
            "description": "Number of students who were rejected"
          },
          "pendingCount": {
            "type": "number",
            "description": "Number of students with pending status"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "description": "Error message"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/admission/upload": {
      "post": {
        "summary": "Upload and process admission data from Excel file",
        "tags": [
          "Admission Data"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file",
                  "period",
                  "year"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Excel file containing admission data"
                  },
                  "period": {
                    "type": "string",
                    "description": "Academic period (e.g., 2025-2)"
                  },
                  "year": {
                    "type": "integer",
                    "description": "Academic year"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "File processed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Admission data processed successfully"
                        },
                        "recordsProcessed": {
                          "type": "integer",
                          "description": "Number of records processed"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Missing file or parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/admission/": {
      "get": {
        "summary": "Get admission data with optional filters",
        "tags": [
          "Admission Data"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "period",
            "schema": {
              "type": "string"
            },
            "description": "Filter by academic period (e.g., 2025-2)"
          },
          {
            "in": "query",
            "name": "year",
            "schema": {
              "type": "integer"
            },
            "description": "Filter by academic year"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            },
            "description": "Number of records per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Admission data retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "records": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/AdmissionData"
                          }
                        },
                        "pagination": {
                          "type": "object",
                          "properties": {
                            "page": {
                              "type": "integer"
                            },
                            "limit": {
                              "type": "integer"
                            },
                            "total": {
                              "type": "integer"
                            },
                            "pages": {
                              "type": "integer"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/admission/stats": {
      "get": {
        "summary": "Get admission statistics",
        "tags": [
          "Admission Data"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "period",
            "schema": {
              "type": "string"
            },
            "description": "Filter by academic period (e.g., 2025-2)"
          },
          {
            "in": "query",
            "name": "year",
            "schema": {
              "type": "integer"
            },
            "description": "Filter by academic year"
          }
        ],
        "responses": {
          "200": {
            "description": "Admission statistics retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/AdmissionStats"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}